% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailtoR.R
\name{mailtoR}
\alias{mailtoR}
\title{Create a friendly user interface for sending emails}
\usage{
mailtoR(email, text, subject = NULL, cc = NULL, bcc = NULL, body = NULL)
}
\arguments{
\item{email}{the emails of the recipients}

\item{text}{the link text that will be displayed on the ui}

\item{subject}{the subject of the email. Defaults to NULL}

\item{cc}{the emails of the CC (carbon copy) recipients. Defaults to NULL}

\item{bcc}{the emails of the BCC (blind carbon copy) recipients. Defaults to NULL}

\item{body}{the body of the email. Defaults to NULL}
}
\value{
a user interface for sending emails
}
\description{
Create a friendly user interface for sending emails
}
\examples{


if (interactive()) {



ui <- fluidPage(

    mailtoR(email = "",
    text = "click here to send an email"),
    use_mailtoR()

    )

server <- function(input, output){


}


shinyApp(ui, server)



}
}
