% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mailtoR.R
\name{use_mailtoR}
\alias{use_mailtoR}
\title{Enable the MailtoUI javascript library}
\usage{
use_mailtoR()
}
\value{
called for the side effect of activating the MailtoUI library
}
\description{
The function activates the capabilities of the MailtoUI javascript library. The user needs to set it at the end of the shiny ui.
}
\examples{
\donttest{

# Put the function at the bottom of the ui

use_mailtoR()



}
}
